cleanup disasm_wrapper & fix pointer size

+++ src/disasm_wrapper.c
@@ -213,7 +213,7 @@ intercept_disasm_next_instruction(struct intercept_disasm_context *context,
 	struct intercept_disasm_result result = {.address = code, 0, };
 	const unsigned char *start = code;
 	size_t size = (size_t)(context->end - code + 1);
-	uint64_t address = (uint64_t)code;
+	uint32_t address = (uint32_t)code;
 
 	if (!cs_disasm_iter(context->handle, &start, &size,
 	    &address, context->insn)) {
@@ -231,12 +231,6 @@ intercept_disasm_next_instruction(struct intercept_disasm_context *context,
 	} else {
 		result.is_nop = false;
 	}
-	// for (unsigned char i = 0; i < context->insn->detail->regs_read_count; i++) {
-		// if (context->insn->detail->regs_read[i] == RISCV_REG_) {
-		// 	result.is_pc_rel = true;
-		// 	break;
-		// }
-	// }
 	switch (context->insn->id) {
 		case RISCV_INS_AUIPC:
 		case RISCV_INS_JAL:
@@ -254,85 +248,7 @@ intercept_disasm_next_instruction(struct intercept_disasm_context *context,
 		default:
 		result.is_pc_rel = false;
 		break;
-		// is affected by pc
-		// case RISCV_INS_JALR:
-		// case RISCV_INS_JAL:
-		// case RISCV_INS_BEQ:
-		// case RISCV_INS_BNE:
-		// case RISCV_INS_BLT:
-		// context->insn->detail->regs_read;
-		// RISCV_REG_
 	}
-	// switch (context->insn->id) {
-	// 	case X86_INS_JAE:
-	// 	case X86_INS_JA:
-	// 	case X86_INS_JBE:
-	// 	case X86_INS_JB:
-	// 	case X86_INS_JCXZ:
-	// 	case X86_INS_JECXZ:
-	// 	case X86_INS_JE:
-	// 	case X86_INS_JGE:
-	// 	case X86_INS_JG:
-	// 	case X86_INS_JLE:
-	// 	case X86_INS_JL:
-	// 	case X86_INS_JMP:
-	// 	case X86_INS_JNE:
-	// 	case X86_INS_JNO:
-	// 	case X86_INS_JNP:
-	// 	case X86_INS_JNS:
-	// 	case X86_INS_JO:
-	// 	case X86_INS_JP:
-	// 	case X86_INS_JRCXZ:
-	// 	case X86_INS_JS:
-	// 	case X86_INS_LOOP:
-	// 	case X86_INS_CALL:
-	// 		result.is_jump = true;
-	// 		assert(context->insn->detail->x86.op_count == 1);
-	// 		break;
-	// 	case RISCV_INS_:
-	// 		result.is_nop = true;
-	// 		break;
-	// 	default:
-	// 		result.is_jump = false;
-	// 		break;
-	// }
-
-	// result.has_ip_relative_opr = false;
-
-	/*
-	 * Loop over all operands of the instruction currently being decoded.
-	 * These operands are decoded by capstone, and described in the
-	 * context->insn->detail->x86.operands array.
-	 *
-	 * This operand checking serves multiple purposes:
-	 * The destination of any jumping instruction is found here,
-	 * The instructions using RIP relative addressing are found by this
-	 *  loop, e.g.: mov %rax, 0x36eb55d(%rip)
-	 *
-	 * Any instruction relying on the value of the RIP register can not
-	 * be relocated ( including relative jumps, which naturally also
-	 * rely on the RIP register ).
-	 */
-	// for (uint8_t op_i = 0;
-	//     op_i < context->insn->detail->x86.op_count; ++op_i)
-	// 	check_op(&result, context->insn->detail->x86.operands + op_i,
-	// 	    code);
-
-	// result.is_lea_rip = (context->insn->id == X86_INS_LEA &&
-	// 		result.has_ip_relative_opr);
-
-	// if (result.is_lea_rip) {
-	// 	/*
-	// 	 * Extract the four bits from the encoding, which
-	// 	 * specify the destination register.
-	// 	 */
-
-	// 	/* one bit from the REX prefix */
-	// 	result.arg_register_bits = ((code[0] & 4) << 1);
-
-	// 	/* three bits from the ModRM byte */
-	// 	result.arg_register_bits |= ((code[2] >> 3) & 7);
-	// }
 
 	result.is_set = true;
